[Geekeren] #348 Move mockServer to utils
